update rl pong app
